Reduce the amount of unnecessary refcounting
In a lot of places, we're accidentally taking a strong reference to a
variable, which means that internally, we'll be using a lot of
`g_object_ref()` and `g_object_unref()` for no good reason. Especially
on some hot paths (and large address books), this can make a tiny bit of
a difference in terms of performance.