window: Add null-check for list_pane on startup
To implement search-as-you-type, we make sure to focus the search field
if that wasn't the case already. However, if a user types something wile
Contacts is still starting up, the UI hasn't finished by that time
already, so that will lead us to dereferencing a NULL-pointer.

To fix this nice and easy, let's to do a quick NULL-check before we try
to focus anything.

Fixes: https://gitlab.gnome.org/GNOME/gnome-contacts/-/issues/155